Separated logic into functions
